Updated Deck UI Animations
hopefully, this fixes most of the issues (on thing: I forgot to update data after the card has been moved, and was getting children wrong! One thing I need to fix is that I only need to update one card, the card that was used, and not the other ones!